+++ CHANGELOG.md
@@ -10,6 +10,7 @@ BUG FIXES:
   - Quote keys to StringBlockEditor so keys with spaces, parens, and
     so on work properly.
   - When there is no route to host for SSH, re-establish a new connection.
+  - `vagrant package` once again works, no more nil error. [GH-1423]
 
 ## 1.1.0 (March 14, 2013)